# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "laoshiren-backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/huawei/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Huawei Login",
        "operationId": "huawei_login_api_v1_auth_huawei_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HuaweiLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh Session",
        "operationId": "refresh_session_api_v1_auth_refresh_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Delete Me",
        "operationId": "delete_me_api_v1_me_delete",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/attention": {
      "get": {
        "tags": [
          "attention"
        ],
        "summary": "Get Attention",
        "operationId": "get_attention_api_v1_attention_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttentionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attention/{subject_type}/{subject_id}/feedback": {
      "post": {
        "tags": [
          "attention"
        ],
        "summary": "Record Attention Feedback",
        "operationId": "record_attention_feedback_api_v1_attention__subject_type___subject_id__feedback_post",
        "parameters": [
          {
            "name": "subject_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AttentionSubjectType"
            }
          },
          {
            "name": "subject_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Subject Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttentionFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttentionCandidateResponse": {
        "properties": {
          "subject_type": {
            "$ref": "#/components/schemas/AttentionSubjectType"
          },
          "subject_id": {
            "type": "string",
            "format": "uuid",
            "title": "Subject Id"
          },
          "thing_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thing Id"
          },
          "candidate_type": {
            "type": "string",
            "title": "Candidate Type"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "last_surfaced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Surfaced At"
          },
          "next_eligible_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Eligible At"
          },
          "acknowledged": {
            "type": "boolean",
            "title": "Acknowledged"
          }
        },
        "type": "object",
        "required": [
          "subject_type",
          "subject_id",
          "thing_id",
          "candidate_type",
          "severity",
          "summary",
          "due_at",
          "last_surfaced_at",
          "next_eligible_at",
          "acknowledged"
        ],
        "title": "AttentionCandidateResponse"
      },
      "AttentionFeedbackAction": {
        "type": "string",
        "enum": [
          "SURFACED",
          "ACKNOWLEDGED",
          "DISMISSED"
        ],
        "title": "AttentionFeedbackAction"
      },
      "AttentionFeedbackRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/AttentionFeedbackAction"
          },
          "dismissed_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dismissed Until"
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "AttentionFeedbackRequest"
      },
      "AttentionResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttentionCandidateResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AttentionResponse"
      },
      "AttentionSubjectType": {
        "type": "string",
        "enum": [
          "THING",
          "TASK",
          "DEADLINE",
          "BLOCKER"
        ],
        "title": "AttentionSubjectType"
      },
...
      "DevicePlatform": {
        "type": "string",
        "enum": [
          "HARMONYOS",
          "OTHER"
        ],
        "title": "DevicePlatform"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HuaweiLoginRequest": {
        "properties": {
          "id_token": {
            "type": "string",
            "minLength": 1,
            "title": "Id Token"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "platform": {
            "$ref": "#/components/schemas/DevicePlatform",
            "default": "HARMONYOS"
          }
        },
        "type": "object",
        "required": [
          "id_token"
        ],
        "title": "HuaweiLoginRequest"
      },
...
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user_id",
          "expires_at"
        ],
        "title": "LoginResponse"
      },
...
      "UserProfileResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "status",
          "created_at"
        ],
        "title": "UserProfileResponse"
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "DELETING",
          "DELETED"
        ],
        "title": "UserStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}